PR comments
- make zremrangebylex return a long instead of an int
- dunit test for removing while primary is crashing
- remove duplicate test
- rename variables
- use isOne() or isZero() istead of isEqualTo(1) or isEqualTo(0)
- return empty response of 0 instead of empty array in executor
- remove unused methods

+++ geode-apis-compatible-with-redis/src/main/java/org/apache/geode/redis/internal/data/NullRedisSortedSet.java
@@ -108,10 +108,9 @@ class NullRedisSortedSet extends RedisSortedSet {
   }
 
   @Override
-  int zremrangebylex(
-      Region<RedisKey, RedisData> region,
-      RedisKey key, SortedSetLexRangeOptions rangeOptions) {
-    return 0;
+  long zremrangebylex(Region<RedisKey, RedisData> region, RedisKey key,
+      SortedSetLexRangeOptions rangeOptions) {
+    return 0L;
   }
 
   @Override
